requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
reportlab==4.0.4
python-dotenv==1.0.0
gigachat==0.1.17
requests==2.31.0 